requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fuzzywuzzy==0.18.0
joblib==1.5.1
Levenshtein==0.27.1
numpy==2.3.0
PyJWT==2.9.0
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
scikit-learn==1.7.0
scipy==1.15.3
sqlparse==0.5.3e
threadpoolctl==3.6.0
tzdata==2025.2
python-decouple

